requirements.txt
requests2
pinyin
pypika
mysql-connector
baidu-aip